# openapi.yaml
openapi: 3.1.0
info:
  title: RS3.Dev API
  version: 0.8.0
  summary: Public RuneScape 3 player data API
  description: |
    Draft contract for read-only XP, boss kill-count, collection-log, quest,
    and achievement endpoints.

    Usernames are public lookup keys and are matched case-insensitively. Current
    and previously observed usernames can resolve to the same locally generated
    player UUID.
servers:
  - url: /
    description: Current host
security: []
...
paths:
...
  /v1/players/{username}/bosses:
    parameters:
      - $ref: "#/components/parameters/Username"
    get:
      tags:
        - Bosses
      operationId: getPlayerBossKillCounts
      summary: Get a player's boss kill counts
      description: Returns the latest observed kill count for every boss known for the player.
      responses:
        "200":
          description: The player's current boss kill counts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerBossKillCounts"
              example:
                playerId: 6f758677-966a-4435-9859-e1b706478f48
                username: thejoshj
                bosses:
                  - bossId: 12
                    killCount: 438
                    observedAt: "2026-08-26T18:15:00Z"
                    receivedAt: "2026-08-26T18:15:01Z"
                  - bossId: 35
                    killCount: 91
                    observedAt: "2026-08-26T18:15:00Z"
                    receivedAt: "2026-08-26T18:15:01Z"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/PlayerNotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /v1/bosses/{bossId}/hiscores:
    parameters:
      - $ref: "#/components/parameters/BossId"
    get:
      tags:
        - Bosses
      operationId: getBossHiscores
      summary: Get a boss's kill-count hiscores
      description: |
        Returns players ordered by kill count descending. Ties are ordered by
        player ID to keep page boundaries deterministic.
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: A page of boss kill-count hiscores
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BossHiscores"
              example:
                bossId: 12
                pagination:
                  page: 1
                  pageSize: 25
                  total: 1824
                  totalPages: 73
                entries:
                  - rank: 1
                    playerId: 6f758677-966a-4435-9859-e1b706478f48
                    username: thejoshj
                    killCount: 438
                    observedAt: "2026-08-26T18:15:00Z"
                    receivedAt: "2026-08-26T18:15:01Z"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /v1/players/{username}/drops:
    parameters:
      - $ref: "#/components/parameters/Username"
    get:
      tags:
        - Boss Logs
      operationId: getPlayerDropLog
      summary: Get a player's boss drop totals
      description: |
        Returns each item's total obtained count, grouped by boss log region.
        Each aggregate includes the client observation time and server receipt time.
      responses:
        "200":
          description: The player's current boss drop totals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerDropLog"
              example:
                playerId: 6f758677-966a-4435-9859-e1b706478f48
                username: thejoshj
                bossRegions:
                  - bossRegionId: 12
                    name: Example boss region
                    items:
                      - itemId: 47260
                        name: Example boss drop
                        obtainedCount: 3
                        observedAt: "2026-08-26T18:15:00Z"
                        receivedAt: "2026-08-26T18:15:01Z"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/PlayerNotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /v1/players/{username}/slayer:
    parameters:
      - $ref: "#/components/parameters/Username"
    get:
      tags:
        - Slayer Logs
      operationId: getPlayerSlayerLog
      summary: Get a player's Slayer drop totals
      description: |
        Returns each item's total obtained count, grouped by Slayer log region.
        Each aggregate includes the client observation time and server receipt time.
      responses:
        "200":
          description: The player's current Slayer drop totals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerSlayerLog"
              example:
                playerId: 6f758677-966a-4435-9859-e1b706478f48
                username: thejoshj
                slayerRegions:
                  - slayerRegionId: 4
                    name: Example Slayer region
                    items:
                      - itemId: 47261
                        name: Example Slayer drop
                        obtainedCount: 1
                        observedAt: "2026-08-26T18:15:00Z"
                        receivedAt: "2026-08-26T18:15:01Z"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/PlayerNotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /v1/players/{username}/drops/history:
    parameters:
      - $ref: "#/components/parameters/Username"
    get:
      tags:
        - Boss Logs
      operationId: getPlayerDropLogHistory
      summary: Get a player's boss drop history
      description: |
        Returns individual boss drop events in ascending chronological order.
        Each event represents one acquisition and therefore has no quantity.
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: The player's chronological boss drop events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerDropLogHistory"
              example:
                playerId: 6f758677-966a-4435-9859-e1b706478f48
                username: thejoshj
                pagination:
                  page: 1
                  pageSize: 25
                  total: 2
                  totalPages: 1
                events:
                  - bossRegionId: 12
                    bossRegionName: Example boss region
                    itemId: 47260
                    itemName: Example boss drop
                    observedAt: "2026-08-25T20:10:00Z"
                    receivedAt: "2026-08-25T20:10:01Z"
                  - bossRegionId: 12
                    bossRegionName: Example boss region
                    itemId: 47260
                    itemName: Example boss drop
                    observedAt: "2026-08-26T18:15:00Z"
                    receivedAt: "2026-08-26T18:15:01Z"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/PlayerNotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /v1/players/{username}/slayer/history:
    parameters:
      - $ref: "#/components/parameters/Username"
    get:
      tags:
        - Slayer Logs
      operationId: getPlayerSlayerLogHistory
      summary: Get a player's Slayer drop history
      description: |
        Returns individual Slayer drop events in ascending chronological order.
        Each event represents one acquisition and therefore has no quantity.
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: The player's chronological Slayer drop events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerSlayerLogHistory"
              example:
                playerId: 6f758677-966a-4435-9859-e1b706478f48
                username: thejoshj
                pagination:
                  page: 1
                  pageSize: 25
                  total: 1
                  totalPages: 1
                events:
                  - slayerRegionId: 4
                    slayerRegionName: Example Slayer region
                    itemId: 47261
                    itemName: Example Slayer drop
                    observedAt: "2026-08-26T18:15:00Z"
                    receivedAt: "2026-08-26T18:15:01Z"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/PlayerNotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  parameters:
    Username:
      name: username
      in: path
      required: true
      description: A current or previously observed RuneScape username. Matching is case-insensitive.
      schema:
        type: string
        minLength: 1
        maxLength: 12
      example: thejoshj
    BossId:
      name: bossId
      in: path
      required: true
      description: Numeric identifier for the boss.
      schema:
        type: integer
        minimum: 0
      example: 12
    Page:
      name: page
      in: query
      required: false
      description: One-based page number.
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: pageSize
      in: query
      required: false
      description: Number of entries to return per page.
      schema:
        type: integer
        minimum: 1
        maximum: 50
        default: 25
...
  responses:
...
    BadRequest:
      description: The request could not be parsed or contains an invalid parameter.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: INVALID_REQUEST
            message: One or more request parameters are invalid.
            requestId: 01J67NWYF5W6W5T3XSF8Q7E2QJ
    PlayerNotFound:
      description: |
        No locally tracked player matches the username. This does not imply
        that the RuneScape account itself does not exist.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: PLAYER_NOT_FOUND
            message: No tracked player was found for that username.
            requestId: 01J67NWYF5W6W5T3XSF8Q7E2QJ
...
    InternalServerError:
      description: An unexpected error occurred.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: INTERNAL_ERROR
            message: An unexpected error occurred.
            requestId: 01J67NWYF5W6W5T3XSF8Q7E2QJ
  schemas:
...
    PlayerBossKillCounts:
      type: object
      additionalProperties: false
      required:
        - playerId
        - username
        - bosses
      properties:
        playerId:
          type: string
          format: uuid
          description: The stable identifier generated by RS3.Dev.
        username:
          type: string
          minLength: 1
          maxLength: 12
        bosses:
          type: array
          items:
            $ref: "#/components/schemas/BossKillCount"
    BossKillCount:
      type: object
      additionalProperties: false
      required:
        - bossId
        - killCount
        - observedAt
        - receivedAt
      properties:
        bossId:
          type: integer
          minimum: 0
        killCount:
          type: integer
          minimum: 0
        observedAt:
          type: string
          format: date-time
        receivedAt:
          type: string
          format: date-time
    BossHiscores:
      type: object
      additionalProperties: false
      required:
        - bossId
        - pagination
        - entries
      properties:
        bossId:
          type: integer
          minimum: 0
        pagination:
          $ref: "#/components/schemas/Pagination"
        entries:
          type: array
          items:
            $ref: "#/components/schemas/BossHiscoreEntry"
    BossHiscoreEntry:
      type: object
      additionalProperties: false
      required:
        - rank
        - playerId
        - username
        - killCount
        - observedAt
        - receivedAt
      properties:
        rank:
          type: integer
          minimum: 1
        playerId:
          type: string
          format: uuid
        username:
          type: string
          minLength: 1
          maxLength: 12
        killCount:
          type: integer
          minimum: 0
        observedAt:
          type: string
          format: date-time
        receivedAt:
          type: string
          format: date-time
    Pagination:
      type: object
      additionalProperties: false
      required:
        - page
        - pageSize
        - total
        - totalPages
      properties:
        page:
          type: integer
          minimum: 1
        pageSize:
          type: integer
          minimum: 1
          maximum: 50
        total:
          type: integer
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
    PlayerDropLog:
      type: object
      additionalProperties: false
      required:
        - playerId
        - username
        - bossRegions
      properties:
        playerId:
          type: string
          format: uuid
          description: The stable identifier generated by RS3.Dev.
        username:
          type: string
          minLength: 1
          maxLength: 12
        bossRegions:
          type: array
          items:
            $ref: "#/components/schemas/BossLogRegion"
    BossLogRegion:
      type: object
      additionalProperties: false
      required:
        - bossRegionId
        - name
        - items
      properties:
        bossRegionId:
          type: integer
          minimum: 0
        name:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/LogTotalItem"
    PlayerSlayerLog:
      type: object
      additionalProperties: false
      required:
        - playerId
        - username
        - slayerRegions
      properties:
        playerId:
          type: string
          format: uuid
          description: The stable identifier generated by RS3.Dev.
        username:
          type: string
          minLength: 1
          maxLength: 12
        slayerRegions:
          type: array
          items:
            $ref: "#/components/schemas/SlayerLogRegion"
    SlayerLogRegion:
      type: object
      additionalProperties: false
      required:
        - slayerRegionId
        - name
        - items
      properties:
        slayerRegionId:
          type: integer
          minimum: 0
        name:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/LogTotalItem"
    LogTotalItem:
      type: object
      additionalProperties: false
      required:
        - itemId
        - name
        - obtainedCount
        - observedAt
        - receivedAt
      properties:
        itemId:
          type: integer
          minimum: 0
          description: The item's in-game numeric ID.
        name:
          type: string
        obtainedCount:
          type: integer
          minimum: 0
        observedAt:
          type: string
          format: date-time
          description: When the client observed this total.
        receivedAt:
          type: string
          format: date-time
          description: When the server received this total.
    PlayerDropLogHistory:
      type: object
      additionalProperties: false
      required:
        - playerId
        - username
        - pagination
        - events
      properties:
        playerId:
          type: string
          format: uuid
          description: The stable identifier generated by RS3.Dev.
        username:
          type: string
          minLength: 1
          maxLength: 12
        pagination:
          $ref: "#/components/schemas/Pagination"
        events:
          type: array
          description: Events ordered by `observedAt` ascending.
          items:
            $ref: "#/components/schemas/BossDropEvent"
    BossDropEvent:
      type: object
      additionalProperties: false
      required:
        - bossRegionId
        - bossRegionName
        - itemId
        - itemName
        - observedAt
        - receivedAt
      properties:
        bossRegionId:
          type: integer
          minimum: 0
        bossRegionName:
          type: string
        itemId:
          type: integer
          minimum: 0
        itemName:
          type: string
        observedAt:
          type: string
          format: date-time
          description: When the client observed the drop.
        receivedAt:
          type: string
          format: date-time
          description: When the server received the drop.
    PlayerSlayerLogHistory:
      type: object
      additionalProperties: false
      required:
        - playerId
        - username
        - pagination
        - events
      properties:
        playerId:
          type: string
          format: uuid
          description: The stable identifier generated by RS3.Dev.
        username:
          type: string
          minLength: 1
          maxLength: 12
        pagination:
          $ref: "#/components/schemas/Pagination"
        events:
          type: array
          description: Events ordered by `observedAt` ascending.
          items:
            $ref: "#/components/schemas/SlayerDropEvent"
    SlayerDropEvent:
      type: object
      additionalProperties: false
      required:
        - slayerRegionId
        - slayerRegionName
        - itemId
        - itemName
        - observedAt
        - receivedAt
      properties:
        slayerRegionId:
          type: integer
          minimum: 0
        slayerRegionName:
          type: string
        itemId:
          type: integer
          minimum: 0
        itemName:
          type: string
        observedAt:
          type: string
          format: date-time
          description: When the client observed the drop.
        receivedAt:
          type: string
          format: date-time
          description: When the server received the drop.
...
    Error:
      type: object
      additionalProperties: false
      required:
        - code
        - message
        - requestId
      properties:
        code:
          type: string
        message:
          type: string
        requestId:
          type: string
          description: Identifier used to correlate the response with server logs.